back/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==2.0.5
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-jenkins==0.110.0
djangorestframework==3.8.2
drf-base64==0.9.4
drf-schema-adapter==0.9.57
factory-boy==2.11.1
Faker==0.8.13
Genshi==0.7
Inflector==2.0.12
kombu==4.1.0
levit-report==0.9.17
lxml==4.2.1
Pillow==5.1.0
python-dateutil==2.7.2
pytz==2018.4
relatorio==0.8.0
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
vine==1.1.4
zipstream==1.1.4
